ENH: Reduce warning logs during DICOM import
When ctkDICOMIndexer::addFile was called with non-empty destinationDirectoryName (to indicate that file should be copied) then "Ignoring destinationDirectoryName parameter, just taking it as indication we should copy!" warning message was logged for each file. This flooded the logs and had noticeable performance impact.

Since this warning is for developers, it is not necessary to log it for each file during runtime, but the behavior is documented instead in the header file.