BUG: Python3: Fix python console auto-completion
This commit fixes ctkPythonConsoleCompleter. The failure to check the
value returned by PyObject_GetAttrString caused the error flag to be set
and led to an inconsistent completer state. This was a problem only when
built against Python3.

Instead of explicitly checking the type of value returned by "PyObject_GetAttrString()",
the function "PyErr_Print()" could have been called after the call to
"PyObject_GetAttrString". This would have also cleared the internal error
flag was not cleared and printed the following error:

  TypeError: bad argument type for built-in operation